Add layer data and layer layer type
